jhfmuyghbmbghj

+++ ui/src/main/kotlin/com/sendra/ui/screens/transfer/EnhancedTransferScreen.kt
@@ -394,8 +394,8 @@ private fun ShimmerProgressOverlay() {
                         Color.White.copy(alpha = 0.3f),
                         Color.Transparent
                     ),
-                    startX = { width: Float -> width * translateX },
-                    endX = { width: Float -> width * (translateX + 0.5f) }
+                    startX = { width: Float -> (width * translateX) as Float },
+                    endX = { width: Float -> (width * (translateX + 0.5f)) as Float }
                 )
             )
     )
